seriously, screw the lib directory
I've gone back and forth on this one; but for *now*, my opinion
is that this was a mistake, having all these separate files. I
kept modifying lazy.js directly and then calling `rake compile:lib` and
losing my changes. Or, I would modify the "correct" file, but then
forget to run `rake compile:lib`, and then be confused as to why my
change wasn't taking effect. Just a total mess. So I'm going back to
just sticking with lazy.js as a single file.