let's call this 0.1.1
This is admittedly a pretty arbitrary point to turn into a version.
Thing is, I've been slacking on this library for a little bit; I know
0.1.0 is outdated and has some bugs that are now fixed, and I've also
added some really useful stuff like the ability to wrap NodeLists and
HTMLCollections as sequences. So even though this isn't exactly a
pristine point, it's better than 0.1.0; and whatever, we're still in the
0.x zone so anything goes, right?