version bump
This is another case where my reasons for bumping aren't all that specific.
The main thing is the library has seen a bunch of improvements and in
particular there is more uniformity to how array-, object-, and string-like
sequences all behave now. I expect there are still some things to iron out
but that can wait for 0.2.1 :)