I'm actually removing this from the repo...
I don't think it makes sense there. I will try and (from now on)
update the minified JS file for each actual release, and will
add it as a GitHub release. Otherwise it can be confusing for
someone who assumes lazy.min.js is what they want when in
reality it's typically out of date.