implemented AsyncSequence#find
Some of this is getting pretty messy, docs-wise (as in, I'm not
doing a good job documenting my progress as I go). I am aware of
this but writing code is just where my brain is right now. I
sincerely hope that tomorrow morning I'll be in more of a "Let's
update the API docs!" sort of mood.