screw node 0.6!
In all seriousness, I've had no intention from the beginning of
supporting older browsers; so why would I bother supporting an
unsupported Node version?

Of course, the *main* reason I'm doing this is that it's failing in
Travis because of this: https://github.com/isaacs/npm/issues/4379

Maybe that's a bad reason. In which case let me re-emphasize the first
paragraph.