updated README
- removed vanity performance charts
- de-emphasized performance in general to highlight rather the cool
  things you can *do* with lazy
- improved (I think?) a couple of the examples

In hindsight I think it was a bit silly to pit Lazy against Underscore
and Lodash, not because it isn't true (though it may not be anymore --
both libraries have been updated a bunch since I was really active on
this), but because none of these libraries is likely to be a bottleneck
in any real application, even Underscore which last I checked was the
slowest by far.

If devs *really* need insane low-level performance on data manipulation
they'll perform their own tests and draw their own conclusions.