reverted implementation changes from 420dc5e
This commit introduced some helpful tests but applied a fix that was
redundant with changes I already committed to master (well, technically
my changes were redundant, because I don't pay close enough attention to
things, but that's not important right now).

So basically I want the tests but feel fine sticking with the previous
implementation.